@@ -1,20 +1,39 @@
+import { LoginService } from './security/login/login.service';
+import { NotificationService } from './shared/messages/notification.service';
+import { ErrorHandler, Injectable, Injector, NgZone } from '@angular/core';
 import { HttpErrorResponse } from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
 
-import 'rxjs/add/observable/throw';
+//import { throwError } from 'rxjs/operators';
 
-export class ErrorHandler {
-  static handleError(error: HttpErrorResponse | any) {
-    let errorMessage: string;
+import { NgSelectMultipleOption } from '@angular/forms/src/directives';
 
-    if (error instanceof HttpErrorResponse) {
-      const body = error.error;
-      errorMessage = `Erro ${error.status} ao acessar a URL ${error.url} - ${body}`;
-    } else {
-      errorMessage = error.message ? error.message : error.toString();
+@Injectable()
+export class ApplicationErrorHandler extends ErrorHandler {
+  constructor(
+    private ns: NotificationService,
+    private injector: Injector,
+    private zone: NgZone) {
+    super();
+  }
+
+  handleError(errorResponse: HttpErrorResponse | any) {
+    if (errorResponse instanceof HttpErrorResponse) {
+      const message = errorResponse.error.message;
+      this.zone.run(() => {
+        switch (errorResponse.status) {
+          case 401:
+            this.injector.get(LoginService).handleLogin();
+            break;
+          case 403:
+            this.ns.notify(message || 'Não autorizado');
+            break;
+          case 404:
+            this.ns.notify(message || 'Recurso não encontrado');
+            break;
+        }
+      })
     }
 
-    console.log(errorMessage);
-    return Observable.throw(errorMessage);
+    super.handleError(errorResponse);
   }
 }
