@@ -1,7 +1,8 @@
 import { animate } from '@angular/animations';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
+import { map } from 'rxjs/operators';
 
 import { ShoppingCartService } from 'app/restaurant-detail/shopping-cart/shopping-cart.services';
 import { CartItem } from 'app/restaurant-detail/shopping-cart/cart-item.model';
@@ -13,8 +14,7 @@ import { LoginService } from '../security/login/login.service';
 export class OrderService {
   constructor(
     private cartService: ShoppingCartService,
-    private http: HttpClient,
-    private loginService: LoginService
+    private http: HttpClient
   ) { }
 
   cartItems(): CartItem[] {
@@ -50,11 +50,10 @@ export class OrderService {
   // }
 
   checkOrder(order: Order): Observable<string> {
-    let headers = new HttpHeaders();
-    if (this.loginService.isLoggedIn()) {
-      headers = headers.set('Authorization', `Bearer ${this.loginService.user.accessToken}`);
-    }
-    return this.http.post<Order>(`${MEAT_API}/orders`, order, { headers: headers })
-      .map(order => order.id);
+    return this.http.post<Order>(`${MEAT_API}/orders`, order)
+      .pipe(
+        map(order => order.id)
+      );
   }
+
 }
