@@ -1,13 +1,11 @@
+import { map, switchMap, tap } from 'rxjs/operators';
 import { MenuItem } from './../restaurant-detail/menu-item/menu-item.model';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/operator/switchMap';
 
 import { Restaurant } from './restaurant/restaurant.model';
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpParams } from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
-import { ErrorHandler } from './../app.error-handler';
+import { Observable } from 'rxjs';
+
 
 import { MEAT_API } from './../app.api';
 
