@@ -1,9 +1,9 @@
 import { MEAT_API } from '../../app.api';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import 'rxjs/add/operator/do';
-import 'rxjs/add/operator/filter';
+
+import { tap, switchMap, filter } from 'rxjs/operators';
 
 import { User } from './user.model';
 import { Router, NavigationEnd } from '@angular/router';
@@ -18,15 +18,19 @@ export class LoginService {
   lastUrl: string; // para guardar a ultima url, em caso de login, para o app redirecionar para a ultima url
 
   constructor(private Http: HttpClient, private router: Router) {
-    this.router.events.filter(e => e instanceof NavigationEnd)
-      // .subscribe( (e: NavigationEnd) => console.log(e.url));
+    this.router.events
+      .pipe(
+        filter(e => e instanceof NavigationEnd)
+      )
       .subscribe((e: NavigationEnd) => this.lastUrl = e.url);
   }
 
   login(email: string, password: string): Observable<User> {
     return this.Http
       .post<User>(`${MEAT_API}/login`, { email: email, password: password })
-      .do(user => this.user = user)
+      .pipe(
+        tap(user => this.user = user)
+      );
   }
 
   isLoggedIn(): boolean {
