@@ -1,13 +1,19 @@
+import { ApplicationErrorHandler } from './app.error-handler';
+import { AuthInterceptor } from './security/auth.interceptor';
 import { LeaveOrderGuard } from './order/leave-order.guard';
 import { LoggedInGuard } from './security/loggedin.guard';
 import { LoginService } from './security/login/login.service';
 import { NotificationService } from './shared/messages/notification.service';
 import { CoreModule } from './core/core.module';
+
 import { BrowserModule } from '@angular/platform-browser';
-import { NgModule, LOCALE_ID } from '@angular/core';
+import { NgModule, LOCALE_ID, ErrorHandler } from '@angular/core';
 import { HttpClientModule } from '@angular/common/http';
 import { RouterModule } from '@angular/router';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+
+// necessário para registrar um Interceptor na app
+import { HTTP_INTERCEPTORS } from '@angular/common/http';
 
 import { ROUTES } from './app.routes';
 
@@ -25,7 +31,12 @@ import { MenuItemComponent } from './restaurant-detail/menu-item/menu-item.compo
 import { ReviewsComponent } from './restaurant-detail/reviews/reviews.component';
 import { FormBuilder } from '@angular/forms';
 import { NotFoundComponent } from './not-found/not-found.component';
-import { LocationStrategy, HashLocationStrategy } from '@angular/common';
+
+/** I18N */
+import { LocationStrategy, HashLocationStrategy, registerLocaleData } from '@angular/common';
+import localePt from '@angular/common/locales/pt';
+registerLocaleData(localePt, 'pt');
+
 import { LoginComponent } from './security/login/login.component';
 import { UserDetailComponent } from './header/user-detail/user-detail.component';
 
@@ -62,12 +73,14 @@ import { UserDetailComponent } from './header/user-detail/user-detail.component'
         para entrrar com URL e ele nao se perder (404) */
     // { provide: LocationStrategy, useClass: HashLocationStrategy },
 
-    { provide: LOCALE_ID, useValue: 'pt-BR' },
+    { provide: LOCALE_ID, useValue: 'pt' },
     FormBuilder,
     NotificationService,
     LoginService,
     LoggedInGuard,
-    LeaveOrderGuard
+    LeaveOrderGuard,
+    { provide: HTTP_INTERCEPTORS, useClass: AuthInterceptor, multi: true },
+    { provide: ErrorHandler, useClass: ApplicationErrorHandler }
   ],
   bootstrap: [AppComponent]
 })
