@@ -7,19 +7,20 @@ import { trigger, state, style, transition, animate, keyframes } from '@angular/
 @Component({
   selector: 'mt-shopping-cart',
   templateUrl: './shopping-cart.component.html',
+  preserveWhitespaces: true,
   animations: [
     trigger('row', [
       state('ready', style({ opacity: 1 })),
       transition('void => ready', animate('300ms 0s ease-in', keyframes([
-        style({ opacit: 0, transform: 'translateX(-30px)', offset: 0 }),
-        style({ opacit: 0.8, transform: 'translateX(10px)', offset: 0.8 }),
-        style({ opacit: 1, transform: 'translateX(0px)', offset: 1 }),
+        style({ opacity: 0, transform: 'translateX(-30px)', offset: 0 }),
+        style({ opacity: 0.8, transform: 'translateX(10px)', offset: 0.8 }),
+        style({ opacity: 1, transform: 'translateX(0px)', offset: 1 }),
       ]))),
 
       transition('ready => void', animate('300ms 0s ease-out', keyframes([
-        style({ opacit: 1, transform: 'translateX(0px)', offset: 0 }),
-        style({ opacit: 0.8, transform: 'translateX(-10px)', offset: 0.2 }),
-        style({ opacit: 0, transform: 'translateX(30px)', offset: 1 }),
+        style({ opacity: 1, transform: 'translateX(0px)', offset: 0 }),
+        style({ opacity: 0.8, transform: 'translateX(-10px)', offset: 0.2 }),
+        style({ opacity: 0, transform: 'translateX(30px)', offset: 1 }),
       ])))
     ])
   ]
